Address issues in albalargesorter

+++ AprgAlgorithm/AprgAlgorithm/src/LargeSorter/AlbaLargeSorter.hpp
@@ -90,15 +90,13 @@ private:
             }
             m_blocks.addObjectToBlock(newBlockIterator, objectToSort);
             numberOfObjectsInCurrentBlock++;
-            numberOfObjectsInCurrentBlock = (numberOfObjectsInCurrentBlock < m_configuration.m_minimumNumberOfObjectsPerBlock) ? numberOfObjectsInCurrentBlock : 0;
+            numberOfObjectsInCurrentBlock = (numberOfObjectsInCurrentBlock < static_cast<int>(m_configuration.m_minimumNumberOfObjectsPerBlock)) ? numberOfObjectsInCurrentBlock : 0;
         });
         m_blocks.deleteBlock(blockIterator);
-    }
-    void limitMemoryConsumption()
+    }    void limitMemoryConsumption()
     {
         unsigned int totalMemoryConsumption = calculateTotalMemoryConsumption();
-        transferMemoryBlocksToFileIfNeeded(totalMemoryConsumption);
-    }
+        transferMemoryBlocksToFileIfNeeded(totalMemoryConsumption);    }
     unsigned int calculateTotalMemoryConsumption()
     {
         BlockInformationContainer memoryLimitCache = m_memoryCache.getContainerReference();
